PhysicalFilter: renamed from BusBlocker
This device will soon not be derived from the BusBypass.
